Major (breaking) architecture changes
* Moved files around to classify them better in the 3 packages
* Token has to be provided instead of fetching from the keychain, moved keychain code out of DiscordKit into Swiftcord

+++ Sources/DiscordKit/Gateway/DiscordGateway.swift
@@ -7,7 +7,6 @@
 
 import Foundation
 import os
-import SwiftUI
 import DiscordKitCommon
 import DiscordKitCore
 
@@ -29,7 +28,7 @@ public class DiscordGateway: ObservableObject {
 	public let onAuthFailure = EventDispatch<Void>()
 
     // WebSocket object
-    @Published public var socket: RobustWebSocket!
+    @Published public var socket: RobustWebSocket?
 
     /// A cache for some data received from the Gateway
     ///
@@ -58,24 +57,32 @@ public class DiscordGateway: ObservableObject {
     @Published public var reachable = false
 
     // Logger
-	private let log = Logger(subsystem: Bundle.main.bundleIdentifier ?? DiscordAPI.subsystem, category: "DiscordGateway")
+    private let log = Logger(
+        subsystem: Bundle.main.bundleIdentifier ?? DiscordKitConfig.subsystem,
+        category: "DiscordGateway"
+    )
 
     /// Log out the current user, closing the Gateway socket connection
     ///
     /// This method removes the Discord token from the keychain and
     /// closes the Gateway socket. The socket will _not_ reconnect.
     public func logout() {
+        guard let socket = socket else {
+            log.warning("Cannot log out, socket hasn't been opened. Call connect() first.")
+            return
+        }
+
         log.debug("Logging out on request")
 
         // Remove token from the keychain
-        Keychain.remove(key: "authToken")
+        /*Keychain.remove(key: "authToken")
         // Reset user defaults
         if let bundleID = Bundle.main.bundleIdentifier {
             UserDefaults.standard.removePersistentDomain(forName: bundleID)
-        }
+        }*/
         // Clear cache
         cache = CachedState()
-        objectWillChange.send()
+        cache.objectWillChange.send()
 
         socket.close(code: .normalClosure)
         onAuthFailure.notify()
@@ -83,13 +90,32 @@ public class DiscordGateway: ObservableObject {
 
     /// Opens the socket connection with the Gateway
     ///
-    /// Calls ``RobustWebSocket/open()``
-    ///
     /// > The socket will be automatically opened when an instance of
     /// > ``DiscordGateway`` is created, so there is no need to call
     /// > this method right after initing an instance.
-    public func connect() {
-        socket.open()
+    ///
+    /// - Parameter token: Token to use to authenticate with the Gateway
+    public func connect(token: String) {
+        socket = RobustWebSocket(token: token)
+        evtListenerID = socket!.onEvent.addHandler { [weak self] (t, d) in
+            self?.handleEvent(t, data: d)
+        }
+        authFailureListenerID = socket!.onAuthFailure.addHandler { [weak self] in
+            self?.onAuthFailure.notify()
+        }
+        connStateChangeListenerID = socket!.onConnStateChange.addHandler { [weak self] (c, r) in
+            self?.connected = c
+            self?.reachable = r
+        }
+        socket!.open()
+    }
+
+    public func send(op: GatewayOutgoingOpcodes, data: OutgoingGatewayData) {
+        guard let socket = socket else {
+            log.warning("Not sending data to a non existant socket")
+            return
+        }
+        socket.send(op: op, data: data)
     }
 
     private func handleEvent(_ event: GatewayEvent, data: GatewayData?) {
@@ -135,6 +161,7 @@ public class DiscordGateway: ObservableObject {
                 break
         }
 
+        cache.objectWillChange.send()
         onEvent.notify(event: (event, data))
         log.info("Dispatched event <\(event.rawValue, privacy: .public)>")
     }
@@ -142,40 +169,22 @@ public class DiscordGateway: ObservableObject {
     /// Inits an instance of ``DiscordGateway``
     ///
     /// Refer to ``RobustWebSocket/init()`` for more details about parameters
-    ///
-    /// - Parameters:
-    ///   - connectionTimeout: The timeout before the connection attempt
-    ///   is aborted. The socket will attempt to reconnect if connection times out.
-    ///   - maxMissedACK: Does not have any effect, included for backward
-    ///   compatibility. The current implementation follows the official
-    ///   Discord client for the missed heartbeat ACK tolerance.
-	public init(connectionTimeout: Double = 5, maxMissedACK: Int = 3) {
-        socket = RobustWebSocket()
-        evtListenerID = socket.onEvent.addHandler { [weak self] (t, d) in
-            self?.handleEvent(t, data: d)
+    public init() {
+    }
+
+    private func removeHandlers() {
+        if let evtListenerID = evtListenerID {
+            _ = socket?.onEvent.removeHandler(handler: evtListenerID)
         }
-        authFailureListenerID = socket.onAuthFailure.addHandler { [weak self] in
-            self?.onAuthFailure.notify()
+        if let authFailureListenerID = authFailureListenerID {
+            _ = socket?.onAuthFailure.removeHandler(handler: authFailureListenerID)
         }
-        connStateChangeListenerID = socket.onConnStateChange.addHandler { [weak self] (c, r) in
-            withAnimation {
-                self?.connected = c
-                self?.reachable = r
-            }
+        if let connStateChangeListenerID = connStateChangeListenerID {
+            _ = socket?.onConnStateChange.removeHandler(handler: connStateChangeListenerID)
         }
     }
 
-    deinit {
-        if let evtListenerID = evtListenerID {
-            _ = socket.onEvent.removeHandler(handler: evtListenerID)
-        }
-        if let authFailureListenerID = authFailureListenerID {
-            _ = socket.onAuthFailure.removeHandler(handler: authFailureListenerID)
-        }
-        if let connStateChangeListenerID = connStateChangeListenerID {
-            _ = socket.onConnStateChange.removeHandler(handler: connStateChangeListenerID)
-        }
-    }
+    deinit { removeHandlers() }
 
     // MARK: - Private Event Handlers
